indexer: refactor testing module
