+++ change/@apibara-indexer-6b02cd7d-9389-44a1-9d26-667585103372.json
@@ -1,0 +1,7 @@
+{
+  "type": "prerelease",
+  "comment": "indexer: refactor testing module",
+  "packageName": "@apibara/indexer",
+  "email": "francesco@ceccon.me",
+  "dependentChangeType": "patch"
+}
