+++ examples/cli/package.json
@@ -8,7 +8,8 @@
     "dev": "apibara dev",
     "start": "apibara start",
     "lint": "biome check .",
-    "lint:fix": "pnpm lint --write"
+    "lint:fix": "pnpm lint --write",
+    "test": "vitest"
   },
   "keywords": [],
   "author": "",
@@ -16,7 +17,8 @@
   "devDependencies": {
     "@types/better-sqlite3": "^7.6.11",
     "@types/node": "^20.5.2",
-    "typescript": "^5.6.2"
+    "typescript": "^5.6.2",
+    "vitest": "^1.6.0"
   },
   "dependencies": {
     "@apibara/evm": "workspace:*",
