--- packages/indexer/src/hooks/index.ts
@@ -1,2 +1,0 @@
-export * from "./useKVStore";
-export * from "./useSink";
