+++ packages/indexer/src/index.ts
@@ -1,7 +1,3 @@
 export * from "./indexer";
 export * from "./sink";
 export { useIndexerContext } from "./context";
-
-export * from "./plugins";
-export * from "./vcr";
-export * from "./hooks";
