+++ packages/indexer/src/plugins/kv.ts
@@ -102,6 +102,16 @@ export class KVStore {
   }
 }
 
+export type UseKVStoreResult = InstanceType<typeof KVStore>;
+
+export function useKVStore(): UseKVStoreResult {
+  const ctx = useIndexerContext();
+
+  if (!ctx?.kv) throw new Error("KV Plugin is not available in context!");
+
+  return ctx.kv;
+}
+
 const statements = {
   beginTxn: "BEGIN TRANSACTION",
   commitTxn: "COMMIT TRANSACTION",
