+++ packages/indexer/src/sinks/drizzle/drizzle.test.ts
@@ -9,10 +9,9 @@ import { drizzle } from "drizzle-orm/node-postgres";
 import { serial, text } from "drizzle-orm/pg-core";
 import { Client } from "pg";
 import { beforeAll, beforeEach, describe, expect, it } from "vitest";
-import { useSink } from "../../hooks";
 import { run } from "../../indexer";
-import { generateMockMessages } from "../../testing";
-import { getMockIndexer } from "../../testing/indexer";
+import { generateMockMessages, getMockIndexer } from "../../internal/testing";
+import { useSink } from "../../sink";
 import type { Int8Range } from "./Int8Range";
 import { drizzle as drizzleSink } from "./drizzle";
 import { getDrizzleCursor, pgIndexerTable } from "./utils";
